Get rid of Niflow and unpin Nibabel (#1043)
* remove imports from niflow and replace with vendored functions

* remove niflow, unpin nibabel and jump to v5

* fix unit test

* remove upper bound constraint on nipype version and force higher than 1.8.3 where trackvis was removed

* fix broken import

* fix couple things in function center_nifti_origin

* Add error handling to caller of center_nifti_origin

* remove calls to deprecated functions in vendored code

* missed one get_data...

* fix bad formatting of enum values

* do not raise

* small fix in dwi

+++ test/unittests/utils/test_image.py
@@ -149,7 +149,7 @@ def test_remove_dummy_dimension_from_image(tmp_path):
     from clinica.utils.image import remove_dummy_dimension_from_image
 
     input_data = np.random.randint(low=0, high=10, size=(16, 10, 6, 3))
-    input_image = nib.Nifti1Image(input_data, affine=np.eye(4))
+    input_image = nib.Nifti1Image(input_data.astype(np.int32), affine=np.eye(4))
     nib.save(input_image, tmp_path / "input_image.nii.gz")
 
     result = remove_dummy_dimension_from_image(
